refactor: moved Extracted slots and GroupSlot into base_slots.py
